Inventario: importar deja de tardar minutos y arregla el 500 al cerrar la toma
Importar la planilla mandaba un UPDATE por artículo: 1704 sentencias para
3162 filas. Contra la base remota de Render cada una es una ida y vuelta por
la red, así que la importación diaria tardaba minutos.

La causa es que al guardar, SQLAlchemy agrupa los UPDATE por el conjunto de
columnas que cambió en cada fila, y aquí cada artículo cambiaba columnas
distintas (el nombre sólo si venía, el costo sólo si venía), fragmentándose en
miles de sentencias. Ahora el valor final se calcula en Python y se manda
siempre el mismo juego de columnas, de modo que se resuelve con executemany:
1704 sentencias pasan a 13. Marcar los artículos ausentes también se hacía
recorriéndolos uno por uno y ahora es una sentencia por tanda.

Las reglas se mantienen intactas: la planilla no pisa con vacío lo que ya
había, Defontana no reemplaza el nombre que puso QMS, la ubicación escrita a
mano se respeta y el stock de lo ya contado se sigue congelando.

Aparte, cerrar una toma fallaba con "integer out of range" cuando la
valorización total pasaba de 2.147 millones de pesos, porque los tres totales
eran Integer. Pasan a BigInteger, con su migración.

Incluye pruebas que fijan un techo de consultas para que la lentitud no vuelva
a colarse, y una que comprueba el tipo de las columnas porque las pruebas
corren en SQLite, que no respeta el límite del Integer.

+++ app/models/conteo_inventario.py
@@ -227,9 +227,12 @@ class TomaInventario(db.Model):
     dif_stock = db.Column(db.Integer, default=0, nullable=False)
     dif_costo = db.Column(db.Integer, default=0, nullable=False)
     dif_unidad = db.Column(db.Integer, default=0, nullable=False)
-    valor_qms_total = db.Column(db.Integer, default=0, nullable=False)
-    valor_defontana_total = db.Column(db.Integer, default=0, nullable=False)
-    valor_fisico_total = db.Column(db.Integer, default=0, nullable=False)
+    # BigInteger y no Integer: son la valorización de TODO el inventario en pesos.
+    # El entero normal de PostgreSQL llega hasta 2.147.483.647, o sea unos 2.147
+    # millones; pasado eso, cerrar la toma reventaba con "integer out of range".
+    valor_qms_total = db.Column(db.BigInteger, default=0, nullable=False)
+    valor_defontana_total = db.Column(db.BigInteger, default=0, nullable=False)
+    valor_fisico_total = db.Column(db.BigInteger, default=0, nullable=False)
 
     cerrado_por = db.relationship("Usuario")
     detalles = db.relationship(
